Remove shared convert script and inline logic
